fix(vercel deployement): rename ProductDetailSection folder to productDetailsSection
